fix updater builder
